rename files in order to use the right convention
