+++ src/cmd/logic-function.js
